+++ terraform/variables.tf
@@ -136,6 +136,10 @@ variable "policy_assignment_display_name" {
   default = "inspec_policy_assignment_name"
 }
 
+variable "express_route_name" {
+  default = "AARNet"
+}
+
 variable "inspec_db_migration_service" {
   default = {
     name = "inspec-compliance-migration-dev"
@@ -149,4 +153,8 @@ variable "inspec_container_group_name" {
 
 variable "sample_directory_object" {
   default = "adc07321-ef2b-44d5-a210-559aa5f10f2d"
+}
+
+variable "inspec_compliance_redis_cache_name" {
+  default = "inspec-compliance-redis-cache"
 }