+++ CHANGELOG.md
@@ -1,10 +1,30 @@
 # Changelog
-<!-- latest_release 1.58.0 -->
+<!-- latest_release 1.61.0 -->
+## [v1.61.0](https://github.com/inspec/inspec-azure/tree/v1.61.0) (2021-08-13)
+
+#### Merged Pull Requests
+- F/express route [#419](https://github.com/inspec/inspec-azure/pull/419) ([sa-progress](https://github.com/sa-progress))
+<!-- latest_release -->
+
+## [v1.60.0](https://github.com/inspec/inspec-azure/tree/v1.60.0) (2021-08-13)
+
+#### Merged Pull Requests
+- F/azure dns [#416](https://github.com/inspec/inspec-azure/pull/416) ([sa-progress](https://github.com/sa-progress))
+
+## [v1.59.0](https://github.com/inspec/inspec-azure/tree/v1.59.0) (2021-08-12)
+
+#### Merged Pull Requests
+- Support azure redis cache [#394](https://github.com/inspec/inspec-azure/pull/394) ([sathish-progress](https://github.com/sathish-progress))
+
+## [v1.58.1](https://github.com/inspec/inspec-azure/tree/v1.58.1) (2021-08-11)
+
+#### Merged Pull Requests
+- Add .DS_Store to gitignore [#428](https://github.com/inspec/inspec-azure/pull/428) ([IanMadd](https://github.com/IanMadd))
+
 ## [v1.58.0](https://github.com/inspec/inspec-azure/tree/v1.58.0) (2021-08-06)
 
 #### Merged Pull Requests
 - F/bastion hosts [#414](https://github.com/inspec/inspec-azure/pull/414) ([sa-progress](https://github.com/sa-progress))
-<!-- latest_release -->
 
 ## [v1.57.0](https://github.com/inspec/inspec-azure/tree/v1.57.0) (2021-08-06)